Added Testing cog
The testing cog is used for testing out the functionality of python and discord.py and is not meant to be running in the bot for production.

In the cog, i've added an event where the bot posts to discord when a user reacts to any message with an emoji.