Meetupmap and status commands
The meetupmap command will reply with a thumbnail of the meetup map, a link to the map, and directions on how to add a pin to the map.
The status command allows you to set the bots playing status on the member sidebar. using status clear will clear it. It will always add the help command to the end of the status.