Finished media move command, non-media messages now auto del in media channels, updated stickers
any message in a channel that's name ends with "media" is scanned. Move reactions and auto deletion only apply to those channels. If a message is posted by an admin, it will not be deleted. If cookie dough does not detect an embed or attachment, she will guess a message is not media and delete it in media channels.