"move to general" bugfix, added wetting gif to stickers
renamed the reactor variable to "ractor_id", added "reactor_member" variable. This was done so that reactor.color.value functioned properly, as that information can only be pulled by having all member information, not just the member ID.

added "if not reaction.custom_emoji: /n continue", so now the "movetogeneral" emoji reaction will be properly removed if it's later in the list than a non-custom emoji.